+++ iblrig/hardware_validation.py
@@ -496,6 +496,19 @@ class ValidatorCamera(Validator):
                 yield Result(Status.WARN, f'Measured TTL rate: {trigger_rate:.1f} Hz (expecting {target_rate} Hz)')
         return True
 
+class ValidatorDataFolders(Validator):
+    _name = 'Data Folders'
+
+    def _run(self):
+        if self.iblrig_settings.iblrig_remote_data_path is None:
+            yield Result(Status.SKIP, 'iblrig_remote_data_path has not been set in hardware_settings.yaml - skipping validation')
+            return False
+        elif self.iblrig_settings.iblrig_remote_data_path.exists():
+            yield Result(Status.PASS, f"Remote data path '{self.iblrig_settings.iblrig_remote_data_path}' is accessible")
+            return True
+        else:
+            yield Result(Status.FAIL, f"Cannot access remote data path '{self.iblrig_settings.iblrig_remote_data_path}'",
+                solution='Check network connection and mapping of network drive')
 
 class ValidatorAlyx(Validator):
     _name = 'Alyx'